users can only edit/delet their own posts

+++ backend/routes/posts.js
@@ -73,9 +73,13 @@ router.put(
       content: req.body.content,
       imagePath: imagePath,
     });
-    Post.updateOne({ _id: req.params.id }, post).then((result) => {
+    Post.updateOne({ _id: req.params.id, creator: req.userData.userId }, post).then((result) => {
       console.log(result);
-      res.status(200).json({ message: "update successful" });
+      if (result.nModified > 0) { //if the post as actually been updated
+        res.status(200).json({ message: "update successful" });
+      } else {
+        res.status(401).json({ message: "Not Authorized" });
+      }
     });
   }
 );
@@ -118,9 +122,13 @@ router.get("/:id", (req, res, next) => {
 
 //delete posts
 router.delete("/:id", checkAuth, (req, res, next) => {
-  Post.deleteOne({ _id: req.params.id }).then((result) => {
+  Post.deleteOne({ _id: req.params.id, creator: req.userData.userId }).then((result) => {
     console.log(result);
-    res.status(200).json({ message: "Post deleted!" });
+    if (result.n > 0){
+      res.status(200).json({ message: "Post deleted!" });
+    } else {
+      res.status(401).json({ message: "Not Authorized" });
+    }
   });
 });
